Update deploy script

+++ deploy.sh
@@ -1,7 +1,10 @@
 # Manually fix flutter beta web packaging error
 if [[ -d "build/web/assets/assets/images" ]]
 then
-  rm -rf build/web/assets/images
+  if [[ -d "build/web/assets/images" ]]
+  then
+    rm -rf build/web/assets/images
+  fi
   mv build/web/assets/assets/images build/web/assets/
 fi
